% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\docType{data}
\name{aerial.survey}
\alias{aerial.survey}
\title{Simulated bowhead whale aerial survey dataset.}
\format{An mrds data frame (with 2 rows per detection) with 86 observations on the following 11 
variables.
 \describe{
   \item{\code{stratum}:}{ stratum number (a numeric vector).}
   \item{\code{area}:}{ stratum surface area (a numeric vector).}
   \item{\code{transect}:}{ transect number (a numeric vector).}
   \item{\code{L}:}{ transect length (a numeric vector).}
   \item{\code{size}:}{ size of each detected group (NA if no detection).}
   \item{\code{object}:}{ unique identifier for each detection (a numeric vector; NA if no detection).}
   \item{\code{side}:}{ the side of the plane from which each detection was made (NA if no detection).}
   \item{\code{obs}:}{ the observer who made each detection (NA if no detection).}
   \item{\code{bf}:}{ Beaufort sea state at the time of each detection (NA if no detection).}
   \item{\code{x}:}{ perpendicular distances to detections (NA if no detection).}
   \item{\code{y}:}{ perpendicular distances to detections (NA if no detection).}
 }}
\source{
Simulated
}
\usage{
aerial.survey
}
\description{
Simulated data representing an aerial survey of bowhead whales from an aircraft 
flying at 46.3 m/s. These data can used together with the dataset \code{\link{bowhead.hmm.pars}} 
to estimate bowhead whale abundance.
}
\examples{
 data(aerial.survey)
}
