% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{AIC.hmltm}
\alias{AIC.hmltm}
\title{Summarise AIC and related for hmltm object(s).}
\usage{
\method{AIC}{hmltm}(object, ..., sort = TRUE, k = 2, dmax = 10,
  criterion = c("AICc", "AIC"))
}
\arguments{
\item{object}{An object of class "hmltm", as produced by \link{est.hmltm}.}

\item{...}{Other object of class "hmltm".}

\item{sort}{Logical to sort output on \code{criterion} or not.}

\item{k}{Value to multiply sample size by for AIC calculation}

\item{dmax}{Models with \code{citerion} difference values greater than this
(scalar) will be given zero (AIC or AICc) weight}

\item{criterion}{The criterion to use when comparing models and reporting differences:
"AICc", or "AIC".}
}
\value{
Summary of models, number of parameters, log-liklihoods, AIC and AICc values,
difference in AIC or AICc values and associated model weights
}
\description{
Tabulates a summary of AIC and related statistcs for fitted models of class
\code{hmltm}.
}
\details{
Hacked from Murray Efford's \code{AIC.secr} function
}
