% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\docType{data}
\name{beaked.ship}
\alias{beaked.ship}
\title{Beaked whale shipboard survey dataset from Alboran sea.}
\format{A data frame with 81 observations on the following 12 variables.
 \describe{
   \item{\code{stratum}:}{ stratum number (a numeric vector). In this dataset this is a dummy 
   value - just there to have something in this compulsory variable.}
   \item{\code{area}:}{ stratum surface area (a numeric vector). In this dataset this is a dummy 
   value - just there to have something in this compulsory variable.}
   \item{\code{transect}:}{ transect number (a numeric vector). In this dataset this is a dummy 
   value - just there to have something in this compulsory variable.}
   \item{\code{L}:}{ transect length (a numeric vector). In this dataset this is a dummy 
   value - just there to have something in this compulsory variable.}
   \item{\code{x}:}{ perpendicular distances to detections (NA if no detection).}
   \item{\code{y}:}{ perpendicular distances to detections (NA if no detection).}
   \item{\code{size}:}{ size of each detected group (NA if no detection).}
   \item{\code{bf}:}{ Beaufort sea state at the time of each detection (NA if no detection).}
   \item{\code{ht}:}{ Observation platform height in m (NA if no detection).}
   \item{\code{object}:}{ unique identifier for each detection (a numeric vector; NA if no detection).}
 }}
\source{
Ana Canadas.
}
\usage{
beaked.ship
}
\description{
Data from the 2008 & 2009 shipboard survey of bowhead whales in the Alboran sea. Sightings
are real, strata and transects are made up for illustration.
}
\details{
Test dataset that contains only detections (transects without detections have been omitted). 
It is one of the datasets analysed in Borchers et al. (2013).
}
\examples{
 data(beaked.ship)
}
\references{
Borchers, D.L., Zucchini, W., Heide-Jorgenssen, M.P., Canadas, A. and Langrock, R. 2013. 
Using hidden Markov models to deal with availability bias on line transect surveys. Biometrics.
}
