% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{bootstrap.p.with.Et}
\alias{bootstrap.p.with.Et}
\title{Detection probability bootstrap with availability process times.}
\usage{
bootstrap.p.with.Et(dat, pars, hfun, models, survey.pars, hmm.pars,
  control.fit, control.opt, fixed.avail = FALSE, B = 999)
}
\arguments{
\item{dat}{detection data frame constructed by removing all rows with no detections from a 
data frame of the sort passed to \code{\link{est.hmltm}}.}

\item{pars}{starting parameter values, as for \code{\link{est.hmltm}}.}

\item{hfun}{detection hazard function name; same as argument \code{FUN} of \code{\link{est.hmltm}}.}

\item{models}{detection hazard covariate models, as for \code{\link{est.hmltm}}.}

\item{survey.pars}{survey parameters, as for \code{\link{est.hmltm}}.}

\item{hmm.pars}{availability hmm parameters, as for \code{\link{est.hmltm}}. Must have elements
\code{$Et} and \code{$Sigma.Et}}

\item{control.fit}{list controlling fit, as for \code{\link{est.hmltm}}.}

\item{control.opt}{list controlling function \code{\link{optim}}, as for \code{\link{est.hmltm}}.}

\item{fixed.avail}{if TRUE, hmm.pars is treated as fixed, else element \code{$Et} is parametrically 
resampled.}

\item{B}{number of bootstrap replicates.}
}
\value{
A list with the following elements:
\itemize{
\item{callist:}{ input reflection: everything passed to the function, bundled into a list}
\item{bs:}{ a list containing (a) a Bxn matrix \code{$phats} in which each row is the estimated 
detection probabilities for each of the n bootstrapped detections, (b) a Bxn matrix \code{$pars} 
in which each row is the estimated detection hazard parameters, (c) the following vectors 
of length B with estimates from each bootstrap: \code{$p0} (mean estimated p(0) over all 
detections), \code{$phat} (mean estimated detection probability over all detections), and (d) 
a Bx2 matrix \code{$b.Et} in which each row is the mean times unavailable and available.
}
}
}
\description{
Nonparametric bootstrap of detection data with estimation of detection probabilities. If 
\code{fixed.avail}=FALSE, does parametric resampling of mean times available and unavailable for 
every resample of detection data, else treats these mean times as fixed.
}
\details{
The rows of data frame \code{dat} are resampled with replacement to create new data frames with as 
many detections as were in \code{dat}. If \code{fixed.avail}=TRUE, then a pair of new mean times 
available and unavailable (\code{$Et}s) are generated for each resampled data frame, by resampling 
parametrically from a logNormal distribution with mean \code{hmm.pars$Et} and variance-covariance
matrix \code{hmm.pars$Sigma.Et}.

Function \code{\link{fit.hmltm}} is called to estimate detection probabilities and related things
for every bootstrap resample.
}
