% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{bootstrap.p.with.hmm}
\alias{bootstrap.p.with.hmm}
\title{Detection probability bootstrap with availability HMM.}
\usage{
bootstrap.p.with.hmm(dat, pars, hfun, models, survey.pars, hmm.pars.bs,
  control.fit, control.opt, fixed.avail = FALSE, B = 999,
  silent = FALSE)
}
\arguments{
\item{dat}{detection data frame constructed by removing all rows with no detections from a 
data frame of the sort passed to \code{\link{est.hmltm}}.}

\item{pars}{starting parameter values, as for \code{\link{est.hmltm}}.}

\item{hfun}{detection hazard function name; same as argument \code{FUN} of \code{\link{est.hmltm}}.}

\item{models}{detection hazard covariate models, as for \code{\link{est.hmltm}}.}

\item{survey.pars}{survey parameters, as for \code{\link{est.hmltm}}.}

\item{hmm.pars.bs}{multiple sets of availability hmm parameters, as output by 
\code{\link{hmmpars.boot}}.}

\item{control.fit}{list controlling fit, as for \code{\link{est.hmltm}}.}

\item{control.opt}{list controlling function \code{\link{optim}}, as for \code{\link{est.hmltm}}.}

\item{fixed.avail}{if TRUE, hmm.pars is treated as fixed, else element \code{$Et} is parametrically 
resampled.}

\item{B}{number of bootstrap replicates.}

\item{silent}{argument of function \code{\link{try}}, controlling error message reporting.}
}
\value{
A list with the following elements:
\itemize{
\item{callist:}{ input reflection: everything passed to the function, bundled into a list}
\item{bs:}{ a list containing (a) a Bxn matrix \code{$phats} in which each row is the estimated 
detection probabilities for each of the n bootstrapped detections, (b) a Bxn matrix \code{$pars} 
in which each row is the estimated detection hazard parameters, and (c) the following vectors 
of length B with estimates from each bootstrap: \code{$Et} (mean times available and unavailable), 
\code{$p0} (mean estimated p(0) over all detections), \code{$phat} (mean estimated detection
probability over all detections), \code{$convergence} convergence diagnostic from \code{optim}.}
}
}
\description{
Nonparametric bootstrap of detection data with re-estimation of detection probabilities. If 
\code{fixed.avail}=FALSE, does nonparametric resampling of availability HMM parameters contained
in hmm.pars.bs for every resample of detection data.
}
\details{
The rows of data frame \code{dat} are resampled with replacement to create new data frames with as 
many detections as were in \code{dat}. If \code{fixed.avail}=TRUE, then a new set of availability 
HMM parameters is obtainded by sampling iwth replacement from \code{hmm.pars.bs}.

Function \code{\link{est.hmltm}} is called to estimate detection probabilities and related things
for every bootstrap resample.
}
