% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootsum}
\alias{bootsum}
\title{Summarise bootstrap results.}
\usage{
bootsum(
  bests,
  ests = NULL,
  cilevel = 0.95,
  write.csvs = FALSE,
  Dstrat = unlist(unique(dimnames(bests)[1])),
  dir = getwd()
)
}
\arguments{
\item{bests}{output from \code{\link{bs.hmltm}}.}

\item{ests}{hmltm object with points estimates.}

\item{cilevel}{confidence level for confidence intervals by percentile method.}

\item{write.csvs}{if TRUE, writes each output (see Value below) to separate .csv file.}

\item{Dstrat}{vector containing the names of strata in which density is to be estimated 
(must be a subset of \code{unique(dat$stratum)}.)}

\item{dir}{directory to which to write outputs if \code{write.csvs} is TRUE.}
}
\value{
If \code{ests} is NULL, returns a list with elements as follows for each statistic 
in \code{bests}:
\itemize{
 \item{nbad} {number of bad estimates of the statistic in question that were excluded from the 
 summary. (Bad estimates occur, for example, when a bootstrap sample involves no detections and
 the estimation function tries to calculate mean group size.)}
 \item{means} {Bootstrap means of the statistic in question.}
 \item{cv} {Percentage CVs of the statistic in question.}
 \item{se} {SEs of the statistic in question.}
 \item{lower} {Lower \code{cilevel} percentiles of the statistic in question.}
 \item{upper} {Upper \code{cilevel} percentiles of the statistic in question.}
}
If \code{ests} is an object of class 'hmltm', returns a data frame with rows only for every 
stratum with detections (and the total) and columns as follows:
\itemize{
\item{Stratum} stratum number
\item{n} original number of detections
\item{n.L} original encounter rate
\item{CV.n.L} percentage coefficient of variation of encounter rate
\item{N.grp} original estimate of group abundance
\item{CV.N.grp} percentage coefficient of variation of group abundance
\item{N.grp.lo} lower bound of group abundance confidence interval
\item{N.grp.hi} upper bound of group abundance confidence interval
\item{Es} original mean group size estimate
\item{CV.Es} percentage coefficient of variation of mean group size
\item{Es.lo} lower bound of mean group size confidence interval
\item{Es.hi} upper bound of mean group size confidence interval
\item{N} original individual abundance estimate
\item{CV.N} percentage coefficient of individual abundance
\item{N.lo} lower bound of individual abundance confidence interval
\item{N.hi} upper bound of individual abundance confidence interval
}
}
\description{
Produces summarry of bootstrap results
}
