% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\docType{data}
\name{bowhead.adat}
\alias{bowhead.adat}
\title{Greenland bowhead whale availability data time series.}
\format{A list of 8 numeric vectors. Their lengths are 191,048, 73,871, 24,673, 9,385, 37,689, 
47,222, 15,490 and 39,989, respectively.}
\usage{
bowhead.adat
}
\description{
A list with elements \code{$a1} to \code{$a8}. Element \code{$ai} is a vector for whale
i containing 0s and 1s, with 0s indicating that the whale was deeper than 2m and 1s indicating it was not. These 
observations are 1 second apart.
}
