% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\docType{data}
\name{bowhead.hmm.pars.bs}
\alias{bowhead.hmm.pars.bs}
\title{Bootstrapped Greenland bowhead whale availability HMM parameters.}
\format{A matrix in which each row is a set of HMM parameters converted to vector format by 
\code{\link{vectorize.hmmpars}}. Each row can be converted back to the format required by
\code{\link{est.hmltm}} using \code{\link{unvectorize.hmmpars}}}
\usage{
bowhead.hmm.pars.bs
}
\description{
Bootstrapped Hidden Markov model (HMM) parameter estimates for bowhead whale 
availability, obtained using by passing \code{\link{bowhead.hmm.pars}} and 
\code{\link{bowhead.adat}} to \code{\link{hmmpars.boot}}.
}
