% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\docType{data}
\name{bowhead.hmm.pars}
\alias{bowhead.hmm.pars}
\title{Greenland bowhead whale availability HMM parameters.}
\format{A list with the following three elements.
 \describe{
   \item{\code{Pi}:}{ a 2x2x8 array containg the 8 HMM transition probability matrices (one for each 
   tagged whale). States can be interpreted as behavioural states, one of which being a state in 
   which the animal is more likely to be available than when in the other state.}
   \item{\code{pm}:}{ a 2x8 matrix containing the 8 vectors of state-dependent Bernoulli availability 
   parameters. \code{pm[1,i]} is the probability of whale i being available given state 1 (the less 
   available behavoural state), and \code{pm[1,i]} is the probability of whale i being available given 
   state 1 (the more available behavoural state).}
   \item{\code{delta}:}{ a 2x8 matrix containing the stationary distribution of \code{Pi} for each 
   whale. So \code{delta[1,i]} is the probability that whale i is in behavioural state 1 when 
   observation starts, and \code{delta[2,i]} is the probability that it is in behavioural state 2 when 
   observation starts.}
 }
 
 @seealso The depth time series data are in object \code{\link{bowhead.depths}} and the binary 
 presence/absence data obtained from these are in object \code{\link{bowhead.adat}}.}
\source{
Laidre et al. (2007).
}
\usage{
bowhead.hmm.pars
}
\description{
Hidden Markov model (HMM) parameter estimates for bowhead whale availability obtained 
from fitting HMMs (using library \code{\link{HiddenMarkov}}) to data from electronic depth-recording 
tags that were attached to eight bowhead whales from the West Greenland population. The tags 
generated 8 time series of durations between 2.6 and 53 hours, with depths recorded every second 
(see Laidre et al., 2007). Following previous practice (Heide-Jorgensen et al., 2007), animals were 
considered to be available for detection only when within 2 m of the surface. The time series were 
accordingly converted into binary availability time series and HMMs were fitted to these. This 
dataset was used in the analyses of Borchers et al. (2013) - albeit with survey data that had no
forward distances.
}
\examples{
 data(bowhead.hmm.pars)
}
\references{
Borchers, D.L., Zucchini, W., Heide-Jorgenssen, M.P., Canadas, A. and Langrock, R. 2013. 
Using hidden Markov models to deal with availability bias on line transect surveys. Biometrics.

Heide-Jørgensen, M. P., Laidre, K., Borchers, D. L., Samarrra,F., and Stern, H. 2007. Increasing 
abundance of bowhead whales in west greenland. Biology Letters 3, 577–580.

Laidre, K., Heide-Jørgensen, M. P., and Nielsen, T. 2007. Role of bowhead whale as a predator in 
West Rreenland. Marine Ecology Progress Series 346, 285–297.
}
