% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{bsample}
\alias{bsample}
\title{Clone of \code{sample} without weird behaviour for length(x)==1.}
\usage{
bsample(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose, or a positive integer. 
See \link{sample} for details.}

\item{size}{a positive number, the number of items to choose from.}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements of the vector being sampled.}
}
\description{
Identical to \code{sample} but when x is an integer it just returns x rather than an
 integer in 1:x (which is what \code{sample} does).
}
