% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{calc.derived}
\alias{calc.derived}
\title{Calculates a bunch of derived statistics from model.}
\usage{
calc.derived(stat, hmmlt, obs = 1:dim(hmmlt$xy)[1])
}
\arguments{
\item{stat}{the statistic name (character variable).}

\item{hmmlt}{the model (as ouput by \code{\link{est.hmltm}}.}

\item{obs}{indices of rows of \code{hmmlt$xy} (i.e. which observations) to use in calculating 
the statistics.}
}
\description{
Calculates a one of a variety of derived statistics from model (see below).
}
\details{
The following are the options for argument \code{stat}:
\describe{
\item{esw:}{effective strip width estimate.}
\item{invesw:}{inverse effective strip width estimate.}
\item{p0:}{estimated probability of detection at perpendicular distance zero.}
\item{p:}{estimated mean probability of detection.}
\item{invp:}{estimated inverse mean probability of detection.}
}
}
