% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{cdfy.u}
\alias{cdfy.u}
\title{Calculate CDF of forward distance.}
\usage{
cdfy.u(y, x, hfun, b, pm, Pi, delta, ymax, dy, theta.f, theta.b, u)
}
\arguments{
\item{y}{forward distance.}

\item{x}{perpendicular distance.}

\item{hfun}{detection hazard function name.}

\item{b}{likelihood function parameters.}

\item{pm}{state-dependent Bernoulli parameters.}

\item{Pi}{Markov model transition probability matrix.}

\item{delta}{Markov model stationary distribution.}

\item{ymax}{forward distance by which detection hazard has fallen to zero.}

\item{dy}{Markov model distance step size.}

\item{theta.f}{REDUNDANT must = 0.}

\item{theta.b}{REDUNDANT must = 90.}

\item{u}{uniform random variable (scalar).}
}
\description{
Calculates CDF(y, given x)-u for single forward distance y, perpendicular distance x, and 
\code{u~runif(1,0,1)}. Currently only for models with no covariates.
}
