% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{chisq.gof.y}
\alias{chisq.gof.y}
\title{Chi-squared goodness-of-fit in forward dimension.}
\usage{
chisq.gof.y(fit, breaks, nys = 250)
}
\arguments{
\item{breaks}{cutpoints on forward-distance axis defining bins}

\item{nys}{number of y-values at which to calculate function before using 
\code{approxfun} approximate it at arbitrary resolution.}

\item{hmltm}{fitted model, as output by \code{\link{est.hmltm}}}
}
\description{
Calculates Chi-squared goodness-of-fit in forward dimension, plots fit, and returns p-value and other stuff.
Returns a data frame  (\code{goftable}) with columns for start and end of bins, observed and expected
frequencies in the bins, and the Chi-squared statistic for the bin. Also 
returns the Chi-squared statistic(\code{X2stat}), the degrees of freedom of the 
test (\code{df}), and the associated p-value (\code{p.Chisq}).
}
\details{
Uses \code{approxfun} to approximate the forward distance pdf and then uses
\code{integrate} to integrate it within each bin.
}
