% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{cv.avail}
\alias{cv.avail}
\title{Calculate CV of mean time available.}
\usage{
cv.avail(hmm.pars, B = 1000)
}
\arguments{
\item{hmm.pars}{object of class \code{hmm.pars}.}

\item{B}{number of bootstrap replicates to do.}
}
\description{
Calculates CV of mean time available and unavailable from an hmm.pars object.
}
