% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{estable}
\alias{estable}
\title{Tabulate point and interval estimates.}
\usage{
estable(est, bs)
}
\arguments{
\item{est}{output from \code{est.hmltm}.}

\item{bs}{output  from \code{bootsum} using the same model that created \code{est}.}
}
\description{
Combines point estimates from \code{est.hmltm} and bootstrap estimates from \code{bs.hmltm} 
to produce a table ready for insertion into a paper. Does not include any strata with no
detections.
}
