% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{f.plot}
\alias{f.plot}
\title{detection function plotting.}
\usage{
f.plot(hmltm, obs = 1:length(hmltm$hmltm.fit$xy$x), new.ymax = NULL,
  new.pars = NULL, theta.f = 0, theta.b = 90, xrange = c(0,
  max(hmltm$hmltm.fit$xy$x)), yrange = c(0, 1.5 *
  max(hmltm$hmltm.fit$xy$y)), nx = 50, ny = nx,
  xlab = "Perpendicular distance", ylab = "Forward distance",
  type = "contour", nlevels = 20, add = FALSE, col = "black",
  logscale = FALSE, theta = 90, phi = 35,
  zlab = "pdf of first detections", values = FALSE, ...)
}
\arguments{
\item{hmltm}{output from \code{\link{est.hmltm}}}

\item{obs}{row (observation) in data to use in plotting}

\item{new.ymax}{a forward distance beyond which things can't be detected (used to override the 
ymax already in \code{hmltm}).}

\item{new.pars}{model parameter vector  (used to override the pars already in \code{hmltm}).}

\item{theta.f}{REDUNDANT must = 0.}

\item{theta.b}{REDUNDANT must = 90.}

\item{xrange}{range of x-axis.}

\item{yrange}{range of y-axis.}

\item{nx}{number of points on x-axis at which to evaluate detection function.}

\item{ny}{number of points on y-axis at which to evaluate detection function.}

\item{xlab}{x label.}

\item{ylab}{y label.}

\item{type}{"contour", "persp", "image" or "both" (for image and contour).}

\item{nlevels}{number of contour levels.}

\item{add}{If TRUE, adds to current plot, else makes new plot.}

\item{col}{colour of plot.}

\item{logscale}{If TRUE, plots hazard values on log scale.}

\item{theta}{argument for \code{\link{image}}.}

\item{phi}{argument for \code{\link{image}}.}

\item{zlab}{label in z-dimension for persp plots}

\item{values}{If TRUE, returns hazard function values on nx by ny grid.}

\item{...}{other arguments to \code{\link{image}}, \code{\link{contour}} or \code{\link{persp}}.}
}
\description{
Plots the detection function (1- the survivor function) of a fitted model in two dimensions.
}
