% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{fit.hmltm}
\alias{fit.hmltm}
\title{Fit detection hazard with a Markov or hidden Markov availability model.}
\usage{
fit.hmltm(xy, pars, FUN, models = list(y = NULL, x = NULL), survey.pars,
  hmm.pars, control.fit, control.optim, groupfromy = NULL)
}
\arguments{
\item{xy}{data frame with one line per detection containing perpendicular distance ($x) and
forward distance ($y) and any covariates in the model.}

\item{pars}{starting parameter values.}

\item{FUN}{detection hazard functional form name (character)}

\item{models}{list of characters with elements \code{$y} and \code{$x} specifying y- and 
x-covariate models. Either \code{NULL} or regression model format (without response on left).}

\item{survey.pars}{list.}

\item{hmm.pars}{HMM parameters of animals' availability data.}

\item{control.fit}{list with elements:
\code{$hessian (logical)} - if TRUE Hessian is estimated and returned, else not;
\code{$nx (scalar)} - determines the number of intervals to use with Simpson's rule 
integration over y. \code{nx=64} seems safe; smaller number makes computing faster.}

\item{control.optim}{as required by \code{\link{optim}}.}

\item{groupfromy}{a forward distance (y) below which all y's are grouped into a single
interval in the likelihood function (i.e. exact y,s < groupfromy are combined into
an interval rather than passed as exact distances).}
}
\value{
A list comprising:
\itemize{
 \item{xy} {dat used in fitting (input reflection).}
 \item{phats} {estimated detection probabilities of all detections.}
 \item{phat} {1/mean(1/phats).}
 \item{pzero} {estimated detection probabilities at perpendicular distance zero for 
 detected population.}
 \item{h.fun} {=FUN (input reflection).}
 \item{models} {=models (input reflection).}
 \item{fit} {output from \code{\link{fit.xy}}.}
 \item{Loglik} {log-likelihood function at MLE.}
 \item{AIC} {AIC.}
 \item{x} {vector of x-values for plotting perpendicular distance fit.}
 \item{p} {vector of detection function values for plotting perpendicular distance fit.}
 \item{fitpars} {a list containing all the given parameters controlling the fit (survey.pars,hmm.pars,
 control.fit,control.optim).}
}
}
\description{
\code{fit.hmltm} Estimates detection probability from (1) line transect data that includes forward 
detection distances and (2) estimated Markov model or hidden Markov model availability prameters.
}
\references{
Borchers, D.L., Zucchini, W., Heide-Jorgenssen, M.P., Canadas, A. and Langrock, R. 
2013. Using hidden Markov models to deal with availability bias on line transect surveys. Biometrics.
}
