% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{fit.xy}
\alias{fit.xy}
\title{Maximises hidden Markov line transect model likelihood.}
\usage{
fit.xy(pars, xy, FUN, models = list(y = NULL, x = NULL), pm, Pi,
  delta = delta, theta.f = 0, theta.b = 90, W, ymax, dy, nx = 50,
  hessian = FALSE, control = list(trace = 5, reltol = 1e-06, maxit =
  200), groupfromy = NULL)
}
\arguments{
\item{pars}{starting parameter values.}

\item{xy}{data frame with one line per detection containing perpendicular distance ($x) and
forward distance ($y) and any covariates in the model.}

\item{FUN}{detection hazard functional form name (character).}

\item{models}{list of characters with elements \code{$y} and \code{$x} specifying y- and 
x-covariate models. Either \code{NULL} or regression model format (without response on left).}

\item{pm}{is a vector of state-dependent Bernoulli distribution parameters (interpretation: the
probability of being available, given state).}

\item{Pi}{is a Markov model transition matrix governing the transition between states. (Square matrix
with each dimension equal to that of \code{pm}.)}

\item{delta}{is a vector specifying the stationary distribution of the Markov model governed by Pi.}

\item{theta.f}{REDUNDANT parameter determining the max forward angle in view (must=0).}

\item{theta.b}{REDUNDANT parameter determining the max forward angle in view (must=90).}

\item{W}{perpendicular truncation distance for fitting. Must be greater than \code{max(xy$x)}.}

\item{ymax}{maximum forward distance that things could be detected. Must be greater than \code{max(xy$y)}.}

\item{dy}{resolution of forward distances for Markov model (typically observer speed times time step size).}

\item{nx}{NOT SURE - I think it is redundant - CHECK}

\item{hessian}{Logical; if TRUE Hessian matrix is returned}

\item{control}{list controlling \code{\link{optim}} optimisation.}

\item{groupfromy}{a forward distance (y) below which all y's are grouped into a single
interval in the likelihood function (i.e. exact y,s < groupfromy are combined into
an interval rather than passed as exact distances).}
}
\value{
A list comprising \code{\link{optim}} ouptut plus element $par containing the estimated parameters 
 on the same scale as input parameters pars (which are transformed before calling \code{\link{optim}}).
}
\description{
\code{fit.xy} A wrapper function for \code{\link{optim}} to get maximum likelihood estimates of
detection hazard function paramters using forward and perpendicular distance data 
(and any associated covariates), using estimated Markov model or hidden Markov model availability 
prameters.
}
