% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{fitted.esw}
\alias{fitted.esw}
\title{Calculates esw from model.}
\usage{
\method{fitted}{esw}(hmmlt, obs = 1:dim(hmmlt$xy)[1], nx = 100,
  to.x = FALSE, all = FALSE)
}
\description{
Calculates effective strip half-width, W*E[p(x)], where W is actual half-width, from model.
 @param hmmlt output from \code{\link{fit.hmltm}}
 @param obs observations (row numbers of \code{hmmlt$xy}) for which to calculate esw 
 @param nx number of x values to use to implement Simpson's rule in perp dist dimension;
 @param to.x If TRUE integrates only to observed x, else integrates to W
 @param all If TRUE then returns esw for every observation, else returns only that for first obs 
 if there are no covariates; always returns esw for every observation if there are covariates.
}
\details{
Calls \code{\link{hmltm.esw}} to calclate effective stript width (esw) for fitted object \code{hmmlt}.
}
