% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{fitted.esw1}
\alias{fitted.esw1}
\title{Calculates esw from model.}
\usage{
\method{fitted}{esw1}(hmmlt, obs = 1:dim(hmmlt$xy)[1], nx = 100,
  to.x = FALSE, all = FALSE, W = NULL)
}
\arguments{
\item{hmmlt}{output from fit.hmltm()}

\item{obs}{observations (row numbers of hmmlt$xy) for which to calculate esw.}

\item{nx}{number of x values to use to implement Simpson's rule in perp dist dimension.}

\item{to.x}{If TRUE integrates only to observed x, else integrates to W.}

\item{all}{If TRUE then returns esw for every observation, else returns only that for
first obs if there are no covariates; always returns esw for every observation
if there are covariates.}

\item{W}{limit of perp. dist integration. If NULL, uses survey.pars$W.}
}
\description{
Calculates effective strip width from fitted model.
}
\details{
Designed to be called by \code{\link{fitted.invp1}}.
Identical to fitted.esw but returns list instead of numerical scalar or vector. This is to allow 
it to be used in \code{\link{NDest}} for estimating density and abundance.
Calls \code{\link{hmltm.esw}} to calclate effective stript width (esw) for 1 observer for fitted 
object \code{hmmlt}.
}
