% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{fitted.invp}
\alias{fitted.invp}
\title{Calculates E[1/p(x)] from model.}
\usage{
\method{fitted}{invp}(hmmlt, obs = 1:dim(hmmlt$xy)[1], nx = 100)
}
\arguments{
\item{hmmlt}{output from \code{\link{fit.hmltm}}}

\item{obs}{observations (row numbers of \code{hmmlt$xy}) for which to calculate esw}

\item{nx}{number of x-values (perpendicular distance values) to use in calculation.}
}
\description{
Calculates mean over perpendicular distance (x) of inverse of fitted values, 1/p(x) for given 
observations, from model.
}
