% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{fitted.invp1}
\alias{fitted.invp1}
\title{Calculates E[1/p(x)] from model.}
\usage{
\method{fitted}{invp1}(hmmlt, obs = 1:dim(hmmlt$xy)[1], nx = 100,
  W = NULL)
}
\arguments{
\item{hmmlt}{output from \code{\link{fit.hmltm}}}

\item{obs}{observations (row numbers of \code{hmmlt$xy}) for which to calculate esw}

\item{nx}{number of x-values (perpendicular distance values) to use in calculation.}

\item{W}{actual half-width over which to integrate (overrides W in \code{hmmlt}).}
}
\description{
Calculates mean over perpendicular distance (x) of inverse of fitted values, 1/p(x) for given 
observations, from model.
}
\details{
Identical to fitted.invp but returns data frame instead of numerical scalar or vector.
This is to allow it to be used in NDest for estimating density and abundance. (Also has extra 
parameter: W)
}
