% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availutils.R
\name{getruns}
\alias{getruns}
\title{Calculate lengths of all runs in availability.}
\usage{
getruns(avail)
}
\arguments{
\item{avail}{is a vecto of binary data (0s and 1s).}
}
\description{
Calculates lengths of the runs of 0s and 1s in a vector.
}
\details{
Returns a list with components \code{run1} being a vector with 
the lengths of all runs of 1s, in order, and  \code{run0} being a vector with 
the lengths of all runs of 0s, in order.
}
