% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{gety.obs}
\alias{gety.obs}
\title{Return all (discrete) forward distances in veiw.}
\usage{
gety.obs(Ix, Inull_yobs, Iyobs, Itheta_f, Itheta_b, Iymax, Idy)
}
\arguments{
\item{Ix}{Perpendicular distance.}

\item{Inull_yobs}{REDUNDANT; must = NULL.}

\item{Iyobs}{Forward distance.}

\item{Itheta_f}{REDUNDANT; must = 0}

\item{Itheta_b}{REDUNDANT; must = 90}

\item{Iymax}{Maximum forward distance in view.}

\item{Idy}{Forward distance step increment.}
}
\description{
C++ function to return all (discrete) forward distances in veiw.
}
