% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{histline}
\alias{histline}
\title{Draws histogram.}
\usage{
histline(height, breaks, lineonly = FALSE, outline = FALSE,
  fill = FALSE, ylim = range(height), xlab = "x", ylab = "y", ...)
}
\arguments{
\item{height}{Height of histogram bars.}

\item{breaks}{Locations of boundaries of histogram bins (must be 1 longer than \code{height}).}

\item{lineonly}{If TRUE, uses \code{\link{lines}} to draw lines on current plot; else uses 
\code{\link{plot}} to draw lines on new plot.}

\item{outline}{If TRUE, draws only the outline (profile) of the histogram; else draws each 
complete bar.}

\item{fill}{If TRUE, uses polygon() to fill barsl in this case valid arguments to polygon() 
are passed via argument(s) "...". If fill==FALSE, valid arguments to plot() or lines() are 
passed via argument(s) "..."}

\item{ylim}{Range of y-axis.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{...}{See aargument \code{fill}.}
}
\description{
Utility function to draw histograms with more options than \code{hist} allows.
}
