% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{hmltm.esw}
\alias{hmltm.esw}
\title{Calculates esw from model.}
\usage{
hmltm.esw(pars, hfun, models, cov, survey.pars, hmm.pars, nx = 100,
  type = "response", to.x = FALSE)
}
\arguments{
\item{pars}{starting parameter values, as for \code{\link{est.hmltm}}.}

\item{hfun}{detection hazard function name; same as argument \code{FUN} of \code{\link{est.hmltm}}.}

\item{models}{detection hazard covariate models, as for \code{\link{est.hmltm}}.}

\item{cov}{covariate matrix with each row corresponding to an observation. Must contain columns 
with variables appearing in \code{models}, and named accordingly, as well as column of perpendicular
distance, named "x". (Perpendicular distances only used if \code{to.x} is TRUE.)}

\item{survey.pars}{survey parameters, as for \code{\link{est.hmltm}}.}

\item{hmm.pars}{availability hmm parameters, as for \code{\link{est.hmltm}}. Must have elements
\code{$Et} and \code{$Sigma.Et}}

\item{nx}{number of x-values (perpendicular distances) to use in evaluating esw.}

\item{type}{if "link", parameter vector \code{pars} is assumed to be on the link scale, else on 
the natural scale}

\item{to.x}{if TRUE integrates only out to \code{cov$x[i]} for observation i (else integrates to 
\code{survey.pars$W}).}
}
\description{
Calculates effective strip half-width from model.
}
\details{
Returns effective strip half-width (esw) for fitted object hmmlt, integrating 
using Simpson's rule.
}
