% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{hmmpars.boot}
\alias{hmmpars.boot}
\title{Bootstrap availability HMM.}
\usage{
hmmpars.boot(availhmm, adat, animals, seed = NULL, B, printprog = TRUE)
}
\arguments{
\item{availhmm}{list with availability HMM paramters, as for the \code{hmm.pars} parameter of 
\code{\link{est.hmltm}}.}

\item{adat}{list of availability data time series. The ith element of the list must be named 
\code{$ai} and must be a vector of 0s and 1s, with 0 corresponding to being unavailable and 1 to 
being available.}

\item{animals}{a vector of integers indicating which of the elements of \code{adat} are to be
resampled.}

\item{seed}{random number seed (integer).}

\item{B}{number of bootstraps to perform (integer).}

\item{printprog}{if TRUE prints progress through animals as it resamples.}
}
\value{
A matrix of dimension \code{B}xT, where T is as described above.
}
\description{
Bootstraps hidden Markov model (HMM) data from multiple observed availability time series 
(one per animal).
}
\details{
Simulates a new series of availability observations (0s and 1s) using functions \code{dthmm} 
and \code{simulate} from library \code{\link{HiddenMarkov}}, then fits a HMM to these data
using function \code{BaumWelch} from library \code{\link{HiddenMarkov}}. Constructs a new 
hmm.pars object from the fitted HMM parameters. 

Does the above \code{B} times, each time reformtting the hmm.pars object as a vector using 
\code{\link{vectorize.hmmpars}} and then entering this as the next row in a matrix of dimension 
\code{B}xT, where T=\code{3+length(animals)*(nstate^2+nstate*2)} and \code{nstate} is the 
number of states in the HMM.
}
\examples{
data(bowhead.hmm.pars)
data(bowhead.adat)
animals=1:8
bs=hmmpars.boot(bowhead.hmm.pars,bowhead.adat,animals,B=3)

}
