% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{hmmpars.paramtetric.boot}
\alias{hmmpars.paramtetric.boot}
\title{Parameteric bootstrap availability HMM.}
\usage{
hmmpars.paramtetric.boot(availhmm, animals, seed = NULL, B, printprog = TRUE)
}
\arguments{
\item{availhmm}{list with availability HMM parameters, as for the \code{hmm.pars} parameter of 
\code{\link{est.hmltm}}. The list must contain these elements:
\itemize{
\item{Pi:}{ a 3D matrix of Markov chain transition probability matrices, with 
element [,,i] containing the 2D matrix (Pi) for animal i 
(state 1=unavailable, state 2=available).}
\item{pm:}{ a matrix with element [,i] containing the state-dependent Bernoulli 
``success'' parameters (p) for each state (first state=unavailable, 2nd=available)
for animal i.}
\item{delta:}{ a matrix with element [,i] containing the stationary disribution
of the Markov chain (first state=unavailable, 2nd=available).}
\item{vcv:}{ a 3D array with element [,,i] containing the estimated variance-
covariance matrix of the following (in this order): logit(Pi[1,1]),
logit(Pi[2,1]),logit(p[1]),logit(p[2]).}
}}

\item{animals}{a vector of integers indicating which of the elements of \code{adat} are to be
resampled.}

\item{seed}{random number seed (integer).}

\item{B}{number of bootstraps to perform (integer).}

\item{printprog}{if TRUE prints progress through animals as it resamples.}
}
\value{
A matrix of dimension \code{B}xT, where T is as described above.
}
\description{
Bootstraps hidden Markov model (HMM) data from estimated parameters and their
variance-covariance matrix (one of each for each animal).
}
\details{
Resamples HMM paramters assuming asymptotic normality of parameters. Constructs a new 
hmm.pars object from the fitted HMM parameters. 

Does the above \code{B} times, each time reformtting the hmm.pars object as a vector using 
\code{\link{vectorize.hmmpars}} and then entering this as the next row in a matrix of dimension 
\code{B}xT, where T=\code{3+length(animals)*(nstate^2+nstate*2)} and \code{nstate} is the 
number of states in the HMM.
}
