% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{invHessvar}
\alias{invHessvar}
\title{Calculates asymptotic variance using Hessian.}
\usage{
invHessvar(stat, hmmlt, nx4spline = 50, doplot = FALSE)
}
\arguments{
\item{stat}{name of statistic to calculate. Valid statistics are "p0" for estimated probability 
at perpendicular distance zero, "p" for mean estimated detection probability over all perpendicular
distances, "invp" for the inverse of mean estimated detection probability over all perpendicular
distances, "esw" for estimated effective strip width, and "invesw" for estimated inverse of effective 
strip width.}

\item{hmmlt}{object output by \code{\link{fit.hmltm}}.}

\item{nx4spline}{is number of points at which to evaluate function for spline interpolation.}

\item{doplot}{if TRUE, plots the spline approximation to the statistic as a function of each 
parameter.}
}
\description{
Calculates asymptotic variance of a statistic derived from a hsltm fit, using the inverse of the 
Hessian matrix and estimated derivative of the statististic with respect to the parameter vector.
The derivative is approximated using function \code{\link{splinefun}}.
}
\seealso{
\code{\link{hmltm.stat}} for calculation of derived statistics, and 
\code{\link{splinefun}} for approximation of function derivatives.
}
