% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{invtfm}
\alias{invtfm}
\title{Parameter inverse transformation function for all models.}
\usage{
invtfm(b, fun)
}
\arguments{
\item{b}{parameters on transformed scale}

\item{fun}{detection hazard function name (character) - see details below.}
}
\description{
Parameter inverse transformation for all models.
}
\details{
Valid detection hazard function names for detection hazards with certain detection at radial
distance zero are "h.EP1.0", "h.EP1x.0", "h.EP2.0", "h.EP2x.0", "h.IP.0".
Inverse transformations are:
\describe{
\item{h.EP1.0}{exponential}
\item{h.EP1x.0}{exponential}
\item{h.EP2.0}{exponential}
\item{h.EP2x.0}{exponential}
\item{h.IP.0}{exponential}
}
}
