% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{jeffa}
\alias{jeffa}
\title{Laake's availability correction factor calculation for a single 2-state Markov model.}
\usage{
jeffa(Pi, w, spd = NULL, E = NULL)
}
\arguments{
\item{Pi}{is Markov model transition matrix (state 1 is UNavailable).}

\item{w}{is maximum forward distance things can be detected.}

\item{spd}{is observer speed. If NULL w is assumed to be in the units of the Markov chain.}

\item{E}{is mean times unavailable (E[1]) and available (E[2]) (in unist of Markov chain). 
If E is not NULL, it is used in preference to Pi for calculations, else it is ignored.}
}
\description{
Calculates the probability that an animal is availabel at least once whil within detectable
region, using the method of Laake et al. (1997), Eqn (4).
}
\details{
See Laake et al. (1997), Eqn (4), or Borchers et al. (2013) for details of the method.
}
\references{
Borchers, D.L., Zucchini, W., Heide-Jorgenssen, M.P., Canadas, A. and Langrock, R. 2013. 
Using hidden Markov models to deal with availability bias on line transect surveys. Biometrics.

Laake, J., Calambokidis, J., Osmek, S., and Rugh, D. 1997. Probability of detecting harbor 
porpoise from aerial surveys: estimating g(0). Journal of Wildlife Management 61, 63-75.
}
