% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lnci.nmin}
\alias{lnci.nmin}
\title{Lognormal confidence interval with lower bound}
\usage{
lnci.nmin(n,Nhat,cv)
}
\arguments{
\item{n}{observed number of population members.}

\item{Nhat}{point estimate of abundance.}

\item{cv}{coefficient of variation of \code{Nhat}.}
}
\description{
Calculates 95\% confidence interval for abundance estimate \code{Nhat}, based on the assumption that 
 \code{Nhat} is lognormally distributed and that \code{n} population members were observed (so that 
 lower bound of confidence interval can't be below \code{n}.
}
