% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{make.covb}
\alias{make.covb}
\title{Constructs linear predictor.}
\usage{
make.covb(b, FUN, models, dat)
}
\arguments{
\item{b}{parameter vector.}

\item{FUN}{detection hazard function name (character).}

\item{models}{list with two components (\code{$y} and \code{$x}) specifying models for detection 
hazard function scale parameters in forward (y) and perpendicular (x) dimensions. If detection 
hazard function form does not have separate scale parameters in y- and x- dimensions, the model
given in \code{$y} is take as the scale parameter model. \code{$y} and \code{$x} must be either 
\code{NULL} or "~<regspec>", where <regspec> is a regression model spefication (e.g. 
\code{height+weight} or \code{height:weight} or \code{height*weight}, etc.).}

\item{dat}{data frame, which must have columns corresponding to variable names in \code{$y} and
\code{$x}.}
}
\description{
Returns parameter vector on linear predictor scale for hazard function \code{FUN}, using 
\code{model} and data frame \code{dat}. Makes a matrix with each row the parameter values for an 
observation, then concatenates these into a single vector (so can pass as vector to C++ code).
}
