% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{make.onesit}
\alias{make.onesit}
\title{Reduces MRDS data frame to CDS data frame.}
\usage{
make.onesit(dat, prefer = 1)
}
\arguments{
\item{dat}{distance data frame in mrds form. Must have columns 'object' (unique detection 
identifier), 'seen' (binary indicating detecteb by observer or not) and 'y' (forward) 
detection distance.}

\item{prefer}{which of the two observers' data to prefer when forward distances are 
missing/equal must be 1 or 2.}

\item{dat}{MRDS data frame.}
}
\description{
Reduces mark-recapture Distance sampling (MRDS) data frame dat, with two lines per detection, to a 
conventional distance sampling (CDS) data frame with a single line per detection. In the case of 
duplicates ("recaptures"), takes the information from the detection made farthest ahead (i.e. that
with larger \code{y}) With duplicates that have the same \code{y} or neither of which have a 
\code{y}, it chooses according to the parameter \code{prefer}. If only one of the duplicates 
has a \code{y}, it chooses that one.
}
