% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{make.survey.pars}
\alias{make.survey.pars}
\title{Makes survey parameter list (\code{survey.pars}) suitable for passing to 
\code{\link{est.hmltm}}.}
\usage{
make.survey.pars(spd, W, ymax, Wl = 0, dT = 1, theta.f = 0,
  theta.b = 90)
}
\arguments{
\item{spd}{is observer speed.}

\item{W}{is perpendicular right-truncation distance for estimation.}

\item{ymax}{is maximum forward distance for estimation - it should be far enough ahead that it is
not possible to detect anyting beyond \code{ymax}.}

\item{Wl}{is perpendicular left-truncation distance for estimation. (After truncating \code{Wl} is 
subtracted from all perpendicular distances.)}

\item{dT}{is the time step on which the availability hidden Markov model operates.}

\item{theta.f}{is REDUNDANT and should be removed. It must equal 0.}

\item{theta.b}{is REDUNDANT and should be removed. It must equal 90.}
}
\description{
\code{make.survey.pars} just puts a bunch of stuff in a list in a format that \code{\link{est.hmltm}}
 expects.
}
\details{
Packs the above in a list suitable for passing as \code{survey.pars} to 
\code{\link{est.hmltm}}.
}
