% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{makeE}
\alias{makeE}
\title{Returns expected time in state for 2-state Markov transition matrices.}
\usage{
makeE(Pi)
}
\arguments{
\item{Pi}{either a 2x2 Markov transition probability matrix or a 2x2xm array with element [,,i]
being a 2x2 Markov transition probability matrix.}
}
\description{
Returns expected time in state for the m 2-state Markov transition matrices Markov transition 
matrices contained in the matrix or array Pi. If m>1 returns a matrix with column i being the
expected time in state 1 ("unavailable") as its first element and expected time in state 2 
("available") as its second; else returns a vector.
}
\examples{
Ea=c(10,12);Eu=c(20,22)
Pi2=makePi(Eu,Ea) # make a set of transition matrices
makeE(Pi2) # recover c(Eu,Ea)
Pi1=makePi(Eu[1],Ea[1]) # make a single transition matrix
makeE(Pi1) # recover c(Eu,Ea)
}
