% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availutils.R
\name{makePi}
\alias{makePi}
\title{Makes Markov transition matrices from mean times available and unavailable.}
\usage{
makePi(Eu, Ea)
}
\arguments{
\item{Eu}{is a vector containing the mean time(s) animals are UNavailable in one 
available-unavailable cycle.}

\item{Ea}{is  a vector containing the mean time(s) animals are available in one 
available-unavailable cycle (must be same length as Eu).}
}
\description{
Makes Markov transition matrices from mean times available (Ea) and unavailable (Eu). 
If Ea and Eu are vectors of length m (they must be the same length), returns a 2x2xm array 
in which element [,,i] is the ith Markov transition matrix; else returns a single 2x2 matrix.
}
\examples{
Ea=c(10,12);Eu=c(20,22)
makePi(Eu,Ea)
makePi(Eu[1],Ea[1])

}
