% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{Npars.from.lNpars}
\alias{Npars.from.lNpars}
\title{Normal from logNormal parameters.}
\usage{
Npars.from.lNpars(mu, Sigma)
}
\arguments{
\item{mu}{multivariate logNormal distribution mean.}

\item{Sigma}{multivariate logNormal distribution variace-covarice matrix.}
}
\description{
Returns mean and covariance matrix of multivariate normal random variables 
which, when logged generate lognormal random variables with mean mu and 
covariance matrix Sigma.
}
