% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numericfy}
\alias{numericfy}
\title{Converts factor or character column of data frame to integer.}
\usage{
numericfy(dat, column = 1, add = FALSE)
}
\arguments{
\item{dat}{A data frame}

\item{column}{The column number or name of the column to be converted.}

\item{add}{If TRUE a new column is added, named `num<column>', where '<column>'
is the value of the \code{column} argument, else the old column is overwritten.}
}
\description{
Replaces a specified column, which may contain factor or character variables, with
uniquely-numbered integer values. Returns the new data frame and a data frame 
showing which factor corresponds to which integer, in a list.
}
