% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{p.xy1}
\alias{p.xy1}
\title{Calculate probability of detection at (x,y).}
\usage{
p.xy1(Ix, Iy, Ihfun, Ib, Ipcu, IPi, Idelta, Iymax, Idy, Itheta_f, Itheta_b,
  Ially, Icdf)
}
\arguments{
\item{Ix}{Perpendicular distance.}

\item{Iy}{Forward distance.}

\item{Ihfun}{Hazard function name.}

\item{Ib}{Hazard function parameter vector.}

\item{Ipcu}{Bernoulli state-dependent probability parameters.}

\item{IPi}{Markov model transition probability matrix.}

\item{Idelta}{Markov model stationary distribution.}

\item{Iymax}{Maximum forward distance in view.}

\item{Idy}{Forward distance step increment.}

\item{Itheta_f}{REDUNDANT; must = 0}

\item{Itheta_b}{REDUNDANT; must = 90}

\item{Ially}{Flag for whether or not to return probabilities at all forward distances in view.}

\item{Icdf}{Flag for whether or not to return cumulative distribution function in forward dimension.
This differes from specifying Ially=TRUE in that Ially=TRUE calculates the cdf from ymax 
to y=0, whereas Icdf=TRUE calculates the cdf from ymax to y.}
}
\description{
C++ function to calculate probability of detection at (x,y), or if cdf=TRUE, BY (x,y).
}
