% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{parametric.resample.hmmpars}
\alias{parametric.resample.hmmpars}
\title{Parameric resample availability HMM.}
\usage{
parametric.resample.hmmpars(
  availhmm,
  animals,
  Pi.only = TRUE,
  seed = NULL,
  nperow = 10,
  printprog = TRUE
)
}
\arguments{
\item{availhmm}{list with availability HMM paramters, as for the \code{hmm.pars} parameter of 
\code{\link{est.hmltm}}.}

\item{animals}{a vector of integers indicating which of the elements of \code{adat} are to be
resampled (e.g. animals=c(1,3,4) means the time series of animals 1, 3 and 4 only will be used).}

\item{Pi.only}{If TRUE, resamples transition probability matrix (TPM) 
parameters but not state-dependent parameters (which are taken as fixed). The 
reason for this is that with availability datasets in which the state-dependent 
parameters are close to 1 and 0, the full variance-covariance matrix is often 
singular, although the TPM component is not.}

\item{seed}{random number seed (integer).}

\item{nperow}{number of iterations printed on one row if printprog=TRUE.}

\item{printprog}{if TRUE, prints progress through animals as it resamples.}
}
\value{
A list with the same format as the input \code{availhmm}
}
\description{
Resamples a hidden Markov model (HMM) data from fitted variance-covariance
matrices (one per animal).
}
\details{
Resamples HMM paramters assuming asymptotic normality of parameters. 
Constructs a new hmm.pars object from the fitted HMM parameters.
}
