% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{poiss.equiv}
\alias{poiss.equiv}
\title{Constructs a HMM equivalent to a Poisson process.}
\usage{
poiss.equiv(availhmm, zero = 0)
}
\arguments{
\item{availhmm}{availability model list of the sort passed to \code{\link{est.hmltm}}.}

\item{zero}{REDUNDANT (I think - CHECK)}
}
\description{
Returns list with $Pi, $pm and $delta of Poisson with same event rate as in the input HMM object 
 availhmm.
}
