% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\docType{data}
\name{porpoise.hmm.pars}
\alias{porpoise.hmm.pars}
\title{Harbour porpoise availability HMM parameters.}
\format{A list with the following three elements.
 \describe{
   \item{\code{Pi}:}{ a 2x2x7 array containg the 7 HMM transition probability matrices (one for each 
   tagged porpoise). States can be interpreted as behavioural states, one of which being a state in 
   which the animal is more likely to be available than when in the other state.}
   \item{\code{pm}:}{ a 2x7 matrix containing the 8 vectors of state-dependent Bernoulli availability 
   parameters. pm[1,i] is the probability of porpoise i being available given state 1 (the less 
   available behavoural state), and pm[1,i] is the probability of porpoise i being available given 
   state 1 (the more available behavoural state).}
   \item{\code{delta}:}{ the stationary distribution of Pi for each porpoise. So delta[1,i] is the 
   probability that porpoise i is in behavioural state 1 when observation starts, and delta[2,i] is 
   the probability that it is in behavioural state 2 when observation starts.}
 }}
\source{
Westgate et al. (1995).
}
\usage{
porpoise.hmm.pars
}
\description{
Hidden Markov model (HMM) parameter estimates for harbour porpoise availability obtained 
the dive durations and proportons of time available of 7 tagged harbour porpoise, as reported by
Westgate et al. (1995). HMM parameters were obtained using funtion \code{\link{make.hmm.pars.from.Et}}.
}
\examples{
 data(porpoise.hmm.pars)
}
\references{
Westgate, A. J., Read, A. J., Berggren, P., Koopman, H. N., and Gaskin, D. E. 1995. Diving behaviour 
of harbour porpoises, Phocoena phocoena. Canadian Journal of Fisheries and Aquatic Sciences 52, 
1064-1073.
}
