% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{resample.hmmpars}
\alias{resample.hmmpars}
\title{Resample availability HMM.}
\usage{
resample.hmmpars(availhmm, adat, animals, seed = NULL, nperow = 10,
  printprog = TRUE)
}
\arguments{
\item{availhmm}{list with availability HMM paramters, as for the \code{hmm.pars} parameter of 
\code{\link{est.hmltm}}.}

\item{adat}{list of availability data time series. The ith element of the list must be named 
\code{$ai} and must be a vector of 0s and 1s, comprising the ith animal's time series, with 0 
corresponding to being unavailable and 1 to being available.}

\item{animals}{a vector of integers indicating which of the elements of \code{adat} are to be
resampled (e.g. animals=c(1,3,4) means the time series of animals 1, 3 and 4 only will be used).}

\item{seed}{random number seed (integer).}

\item{nperow}{number of iterations printed on one row if printprog=TRUE.}

\item{printprog}{if TRUE, prints progress through animals as it resamples.}
}
\value{
A list with the same format as the input \code{availhmm}
}
\description{
Resamples a hidden Markov model (HMM) data from multiple observed availability time series 
(one per animal).
}
\details{
Simulates a new series of availability observations (0s and 1s) using functions \code{dthmm} 
and \code{simulate} from library \code{\link{HiddenMarkov}}, then fits a HMM to these data
using function \code{BaumWelch} from library \code{\link{HiddenMarkov}}. Constructs a new 
hmm.pars object from the fitted HMM parameters.
}
\examples{
data(bowhead.hmm.pars)
data(bowhead.adat)
animals=1:8
resamp=resample.hmmpars(bowhead.hmm.pars,bowhead.adat,animals)

}
