% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{simest}
\alias{simest}
\title{Simulation test hmltm performance}
\usage{
simest(simethod = "hmm", shfun, spars, ehfun = shfun,
  parstart = spars, survey.pars, shmm.pars, ehmm.pars = shmm.pars,
  animal = NULL, adat = NULL, control.opt, control.fit, nsim = 50,
  report.progress = TRUE, En = 100, doplots = FALSE,
  varest = FALSE, hmmpars.bs = NULL, print.n = FALSE,
  silent = FALSE, nx = 100)
}
\arguments{
\item{simethod}{either "hmm" (for availability simulation with HMM) or "animals" (for availability 
simulation by resampling availability or depth dime series).}

\item{shfun}{detection hazard function name for simulation.}

\item{spars}{detection hazard function parameters for simulation.}

\item{ehfun}{detection hazard function name for estimation.}

\item{parstart}{detection hazard function parameter start vallues for estimation.}

\item{survey.pars}{survey parameters, as for \code{\link{est.hmltm}}.}

\item{shmm.pars}{hmm.pars list (as for \code{\link{est.hmltm}}) to be used in simulating.}

\item{ehmm.pars}{hmm.pars list (as for \code{\link{est.hmltm}}) to be used in estimating.}

\item{animal}{which animals' availability data to use (only need if simethod="animals").}

\item{adat}{availability data (only need if simethod="animals"): list of m>=1 vectors 
containing availability binary time series or depths time series.}

\item{control.opt}{optimization parameters passed to \code{\link{optim}}, as for \code{\link{est.hmltm}}.}

\item{control.fit}{fit parameters, as for \code{\link{est.hmltm}}.}

\item{nsim}{number of simulations to do.}

\item{report.progress}{if TRUE, reports progress as does simulations.}

\item{En}{Mean sample size for simulations.}

\item{doplots}{if TRUE does plots of simulaton results.}

\item{varest}{if TRUE, estimates variance (using Hessian matrix).}

\item{hmmpars.bs}{output from \code{\link{hmmpars.boot}}; required if simethod="hmm" and the 
availability HMM parameters are not to be treated as fixed, in which case the
availability HMM parameters for each simulation are obtained by sampling with replacement from 
\code{hmmpars.bs}. If hmmpars.bs is NULL, the availability HMM parameters are treated as fixed
and equal to ehmm.pars.}

\item{print.n}{if TRUE, prints sample size for each simulation.}

\item{silent}{parameter of \code{\link{try}}, controlling error reporting.}

\item{nx}{number of perpendicular distances intervals to use in evaluating detection 
probability, p(x).}
}
\value{
list with these elements:
\itemize{
\item{N:}{ population size.}
\item{esw:}{ true effective strip half-width.}
\item{p0:}{ true p(0).}
\item{p.:}{ true mean detection probability.}
\item{meann:}{ mean sample size across simulations.}
\item{biasNhat:}{ bias of estimated N.}
\item{biaseswhat:}{ bias of estiamted esw.}
\item{biasp0hat:}{ bias of estimated p(0).}
\item{biasphat:}{ bias of estimated mean detection probability.}
\item{biasinvphat:}{ bias of estimated inverse mean detection probability.}
\item{n:}{ sample sizes.}
\item{parest:}{ parameter estimates from each simulation.}
\item{p0hat:}{ p(0) estimates from each simulation.}
\item{phat:}{ mean detection probability estimates from each simulation.}
\item{invphat:}{ inverse mean detection probability estimates from each simulation.}
\item{Nhat:}{ abunance estimates from each simulation.}
\item{invpse:}{ estimated standard error of inverse mean detection probability estimates from each 
simulation (if varest=TRUE).}
\item{p0se:}{ estimated standard error of p(0) estimates from each simulation (if varest=TRUE).}
}
}
\description{
Simulates line transect survey with stochastic animal availability, estimates detection probability
and related parameters, and reports bias, variance of estimators. Currently only for models 
with no covariates.
}
