% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{simhmltm.w}
\alias{simhmltm.w}
\title{Simulate line transect survey data using availability time series.}
\usage{
simhmltm.w(adat, xmax, ymax, spd, animals, hfun, pars, N, dmax = NULL,
  seed = NULL, poiss = FALSE)
}
\arguments{
\item{adat}{list of m>=1 vectors containing availability binary time series or depths time series.}

\item{xmax}{maximum perpendicular distance to simulate.}

\item{ymax}{maximum forward distance to simulate (must be at or beyond point that detection hazard
function is effectively zero).}

\item{spd}{speed observer is moving.}

\item{animals}{a vector of up to m integers specifying which members of the list \code{adat} are to 
be used to simulate availability. (e.g. \code{animals=c(1,3,5)} says the first, third and fifth time
series in \code{adat} are to be used).}

\item{hfun}{detection hazard function name.}

\item{pars}{detection hazard function parameters.}

\item{N}{number of animals within distance \code{xmax} of the transect line and hence subject to 
being detected (or possibly not).}

\item{dmax}{if not NULL, then it is the depth above which animals are considered to be available, and
\code{adat} is taken to be time series of depths, i.e. all \code{adat[[i]]<=dmax} are taken as times
animals are available. If \code{dmax} is NULL, then adat must be binary and all 1's are taken as times
animals are available.}

\item{seed}{random number seed.}

\item{poiss}{if TRUE, the availability time series are randomly permuted, hence generating the 
equivalent of a Poisson availability process.}
}
\value{
Data frame with two elements: \code{$x} and \code{$y}, being perpendicular and forward distances
of detected animals, respectively.
}
\description{
Simulates line transect survey data in which animal availability is simulated by sampling from
time series of availability data (from tagged animals, for example). Currently only for models 
with no covariates.
}
\details{
Simulates (x,y) observations from a population of N animals using detection
hazard specified by hfun & pars, and resampling from the availability time series in adat for
animal availability.
Survey parameters are in survey.pars.
}
