% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleavail.R
\name{simple.a}
\alias{simple.a}
\title{Simple availability correction factor calculation.}
\usage{
simple.a(hmm.pars)
}
\arguments{
\item{hmm.pars}{is as list with an element \code{$Pi} that is a Markov model transition matrix (state 1 is UNavailable).}

\item{E}{is expected time in each state (state 1 in UNavailable).}
}
\description{
Calculates proportion of time an animal with a Markov availability process is available.
}
