% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{strat.estable}
\alias{strat.estable}
\title{Tabulate bootstrap results.}
\usage{
strat.estable(est, cv)
}
\arguments{
\item{est}{element \code{$point$ests} of output from \code{\link{est.hmltm}}.}

\item{cv}{\code{$cv} element of output from \code{\link{strat.estable}}.}
}
\value{
Returns a data frame with columns as follows:
\item{Stratum}{stratum number}
\item{n}{original number of detections}
\item{n.L}{original encounter rate}
\item{CV.n.L}{percentage coefficient of variation of encounter rate}
\item{N.grp}{original estimate of group abundance}
\item{CV.N.grp}{percentage coefficient of variation of group abundance}
\item{N.grp.lo}{lower bound of group abundance confidence interval}
\item{N.grp.hi}{upper bound of group abundance confidence interval}
\item{E.s}{original mean group size estimate}
\item{CV.E.s}{percentage coefficient of variation of mean group size}
\item{N}{original individual abundance estimate}
\item{CV.N}{percentage coefficient of individual abundance}
\item{N.lo}{lower bound of individual abundance confidence interval}
\item{N.hi}{upper bound of individual abundance confidence interval}
}
\description{
Produces brief summarry of bootstrap results
}
