% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{sumsim}
\alias{sumsim}
\title{Summarise simulation results.}
\usage{
sumsim(simout, brief = FALSE)
}
\arguments{
\item{simout}{output from \code{\link{simest}}.}

\item{brief}{if TRUE, prints briefer summary.}
}
\description{
Summarises (and prints) output from \code{\link{simest}}.
}
