% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{truncdat}
\alias{truncdat}
\title{Data truncation.}
\usage{
truncdat(dat, minval = 0, maxval = NULL, twosit = FALSE,
  colnames = c("stratum", "area", "transect", "L", "x", "obs"))
}
\arguments{
\item{dat}{distance data frame. Must have columns for stratum stratum.area, transect, transect.length, 
<any-observer-level_variable>, and if \code{twosit}==TRUE then object as well. These can 
have any names, but the names must be specified, via argument \code{colnames}.}

\item{twosit}{If TRUE, assumes this is an mrds-type dataset with two lines per detection.}

\item{colnames}{name of columns containing stratum, stratum area, transect, transect length, 
<any-observer-level_variable>, and if \code{twosit}==TRUE then object as well, IN THIS ORDER, 
in a character vector. The default value is 
colnames=c("stratum","area","transect","L","x","obs").}

\item{minvalue}{left-truncation point.}

\item{maxvalue}{right-truncation point.}
}
\description{
Left- and/or right-truncates any variable in data frame dat, inserting effort-only record if 
truncation removes all detections on a transect.
Subtracts left-trunction point off all variable values - after all truncation.
}
