% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{unvectorize.hmmpars}
\alias{unvectorize.hmmpars}
\title{Reformat vector as hmm.pars object.}
\usage{
unvectorize.hmmpars(hv)
}
\arguments{
\item{hv}{vector that was vectorised using \code{\link{vectorize.hmmpars}}.}
}
\value{
A hmm.pars object (a list).
}
\description{
Reformats vector that was vectorised using \code{\link{vectorize.hmmpars}}, as a hmm.pars object.
}
