% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{vectorize.hmmpars}
\alias{vectorize.hmmpars}
\title{Reformat hmm.pars object as a vector.}
\usage{
vectorize.hmmpars(hmmpars)
}
\arguments{
\item{hmmpars}{list with availability HMM paramters, as for the \code{hmm.pars} parameter of 
\code{\link{est.hmltm}}.}
}
\value{
A numeric vector.
}
\description{
Reformats hmm.pars object as a vector - so that it can easily be written to file.
}
