% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{hmmlt.gof.x}
\alias{hmmlt.gof.x}
\title{Goodness-of-fit in perpendicular dimension.}
\usage{
hmmlt.gof.x(hmltm, ks.plot = TRUE)
}
\arguments{
\item{hmltm}{fitted model, as output by \code{\link{est.hmltm}}}

\item{ks.plot}{If TRUE, does Q-Q plot. Point corresponding to largest difference between
empirical and theoretical cdf (on which the Kolmogarov-Smirnov test is based) is circled in red.}
}
\description{
Calculates goodness-of-fit in perpendicular dimension, plots fit, and returns p-value and 
other stuff. Returns two p-values: \code{p.ks} is the Kolmogarov-Smirnov p-value (which is
based on only the largest difference between emprical and theoretical cdfs), and Cramer-von Mises
p-value (which is based on all cdf values).
}
