% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{makePi}
\alias{makePi}
\title{Makes Markov transition matrices from mean times available and unavailable.}
\usage{
makePi(Eu, Ea)
}
\arguments{
\item{Eu}{is the mean time UNavailable in one available-unavailable cycle.}

\item{Ea}{is the mean time available in one available-unavailable cycle.}
}
\description{
Makes Markov transition matrices from mean times available (Ea) and unavailable (Eu). 
If Ea and Eu are vectors of length m (they must be the same length), returns a 2x2xm array 
in which element [,,i] is the ith Markov transition matrix; else returns a single 2x2 matrix.
}
\examples{
Ea=c(10,12);Eu=c(20,22)
makePi(Eu,Ea)
makePi(Eu[1],Ea[1])
}
