% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{NDest}
\alias{NDest}
\title{Estimates density and abundance.}
\usage{
NDest(dat, hmltm.fit, W)
}
\arguments{
\item{dat}{MRDS data frame. Must have cols "stratum","area","transect","L","size","object","x","y" 
(and possibly others).}

\item{hmltm.fit}{output from \code{\link{fit.hmltm}}.}

\item{W}{perpendicular truncation distance for estimation.}
}
\description{
Horvitz-Thompson like estimation of density and abundance of groups and of individuals, as well as
of group size (estimated as the ratio of individual density and group density estimates). Produces 
estimates by stratum and over all strata.
}
