% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{simhmltm}
\alias{simhmltm}
\title{Simulate line transect survey data using availability HMM.}
\usage{
simhmltm(nw, hfun, pars, hmm.pars, survey.pars, print.progress = TRUE)
}
\arguments{
\item{nw}{number of animals to simulate.}

\item{hfun}{detection hazard function name.}

\item{pars}{detection hazard function parameters.}

\item{hmm.pars}{availability hidden Markov model parameters (as per \code{\link{est.hmltm}}).}

\item{survey.pars}{survey parameters, as for \code{\link{est.hmltm}}.}

\item{print.progress}{if TRUE prints progress through simulations.}
}
\value{
Data frame with two elements: \code{$x} and \code{$y}, being perpendicular and forward distances
of detected animals, respectively.
}
\description{
Simulates line transect survey data in which animal availability is simulated by generating
time series of availability data from the hidden Markov model(s) with parameters specified in
argument \code{hmm.pars}. Currently only for models with no covariates.
}
\details{
Simulates (x,y) observations from a population of nw animals using detection
hazard specified by hfun & pars, and HMM availability process specified by hmm.pars.
Survey parameters are in survey.pars.
}
