% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsltm.R
\name{strat.estable}
\alias{strat.estable}
\title{Tabulate bootstrap results.}
\usage{
strat.estable(est, cv)
}
\arguments{
\item{est}{output from \code{\link{est.hmltm}}.}

\item{cv}{\code{$cv} element of output from \code{\link{bs.hmltm}}.}
}
\value{
Returns a data frame with columns as follows: \cr
Stratum: stratum number \cr
n: original number of detections \cr
n.L: original encounter rate \cr
CV.n.L: percentage coefficient of variation of encounter rate\cr
N.grp original estimate of group abundance\cr
CV.N.grp: percentage coefficient of variation of group abundance\cr
N.grp.lo: lower bound of group abundance confidence interval\cr
N.grp.hi: upper bound of group abundance confidence interval\cr
E.s: original mean group size estimate\cr
CV.E.s: percentage coefficient of variation of mean group size\cr
N: original individual abundance estimate\cr
CV.N: percentage coefficient of individual abundance\cr
N.lo: lower bound of individual abundance confidence interval\cr
N.hi: upper bound of individual abundance confidence interval\cr
}
\description{
Produces brief summarry of bootstrap results
}
